Fixed wrong coverage report

+++ src/main/shell/run-sonar.sh
@@ -158,7 +158,13 @@ if [[ "$workspaceFile" != "" ]] ; then
 	xctoolCmdPrefix="xctool -workspace $workspaceFile -sdk iphonesimulator ARCHS=i386 VALID_ARCHS=i386 CURRENT_ARCH=i386 ONLY_ACTIVE_ARCH=NO"
 else
 	xctoolCmdPrefix="xctool -project $projectFile -sdk iphonesimulator ARCHS=i386 VALID_ARCHS=i386 CURRENT_ARCH=i386 ONLY_ACTIVE_ARCH=NO"
-fi	
+fi
+
+# Count projects
+projectCount=$(echo $projectFile | sed -n 1'p' | tr ',' '\n' | wc -l | tr -d '[[:space:]]')
+if [ "$vflag" = "on" ]; then
+    echo "Project count is [$projectCount]"
+fi
 
 # Source directories for .h/.m files
 srcDirs=''; readParameter srcDirs 'sonar.sources'
@@ -207,22 +213,11 @@ if [ "$vflag" = "" -a "$nflag" = "" ]; then
 fi
 
 # Create sonar-reports/ for reports output
-if [[ ! (-d "sonar-reports/build") && ("$nflag" != "on") ]]; then
-	if [ "$vflag" = "on" ]; then
-		echo 'Creating directory sonar-reports/'
-	fi
-	mkdir sonar-reports
-	mkdir sonar-reports/build
-	if [[ $? != 0 ]] ; then
-		stopProgress
-    	exit $?
-	fi
-else
-    if [ "$vflag" = "on" ]; then
-		echo 'Clearing sonar-reports/'
-	fi
-    rm -rf sonar-reports/*
+if [ "$vflag" = "on" ]; then
+    echo 'Creating directory sonar-reports/'
 fi
+rm -rf sonar-reports
+mkdir sonar-reports
 
 # Extracting project information needed later
 echo -n 'Extracting Xcode project information'
@@ -246,25 +241,6 @@ else
 
 	echo -n 'Computing coverage report'
 
-	# We do it for every xcodeproject (in case of workspaces)
-
-	# Extract the path to the .gcno/.gcda coverage files
-	echo $projectFile | sed -n 1'p' | tr ',' '\n' > tmpFileRunSonarSh
-	while read projectName; do
-
-        projectName=$(basename $projectName .xcodeproj)
-        coverageFilesPath=$(grep 'command' compile_commands.json | sed 's#^.*-o \\/#\/#;s#",##' | grep "${projectName%%.*}.build" | awk 'NR<2' | sed 's/\\\//\//g' | sed 's/\\\\//g' | xargs -0 dirname)
-		if [ "$vflag" = "on" ]; then
-			echo
-			echo "Path for .gcno/.gcda coverage files is: $coverageFilesPath"
-		fi
-
-		# Create symlink to avoid gcovr bug with --object-directory
-        ln -s $coverageFilesPath sonar-reports/build/$projectName
-
-	done < tmpFileRunSonarSh
-	rm -rf tmpFileRunSonarSh
-
 	# Build the --exclude flags
     excludedCommandLineFlags=""
     if [ ! -z "$excludedPathsFromCoverage" -a "$excludedPathsFromCoverage" != " " ]; then
@@ -278,8 +254,15 @@ else
         echo "Command line exclusion flags for gcovr is:$excludedCommandLineFlags"
     fi
 
+    # Create symlink on the build directory to enable its access from the workspace
+    coverageFilesPath=$(grep 'command' compile_commands.json | sed 's#^.*-o \\/#\/#;s#",##' | grep "${projectName%%.*}.build" | awk 'NR<2' | sed 's/\\\//\//g' | sed 's/\\\\//g' | xargs -0 dirname)
+    splitIndex=$(awk -v a="$coverageFilesPath" -v b="/Intermediates" 'BEGIN{print index(a,b)}')
+    coverageFilesPath=$(echo ${coverageFilesPath:0:$splitIndex}Intermediates)
+    ln -s $coverageFilesPath sonar-reports/build
+
     # Run gcovr with the right options
     runCommand "sonar-reports/coverage.xml" gcovr -r . $excludedCommandLineFlags --xml
+
 	
 fi	
 
@@ -318,14 +301,43 @@ if [ "$fauxpas" = "on" ]; then
     if [ $? -eq 0 ]; then
 
         #FauxPas
-        echo $projectFile | sed -n 1'p' | tr ',' '\n' > tmpFileRunSonarSh
-	    while read projectName; do
+        echo -n 'Running FauxPas...'
 
-            echo -n 'Running FauxPas...'
-            fauxpas -t $appScheme -o json check $projectName --workspace $workspaceFile --scheme $appScheme > sonar-reports/$(basename $projectName .xcodeproj)-fauxpas.json
+        if [ "$projectCount" = "1" ]
+        then
 
-        done < tmpFileRunSonarSh
-	    rm -rf tmpFileRunSonarSh
+            fauxpas -o json check $projectFile --workspace $workspaceFile --scheme $appScheme > sonar-reports/fauxpas.json
+
+
+        else
+
+            echo $projectFile | sed -n 1'p' | tr ',' '\n' > tmpFileRunSonarSh
+            while read projectName; do
+
+                xcodebuild -list -project $projectName | sed -n '/Schemes/,$p' | while read scheme
+                do
+
+                if [ "$scheme" = "" ]
+                then
+                exit
+                fi
+
+                if [ "$scheme" == "${scheme/Schemes/}" ]
+                then
+                    if [ "$scheme" != "$testScheme" ]
+                    then
+                        projectBaseDir=$(dirname $projectName)
+                        workspaceRelativePath=$(python -c "import os.path; print os.path.relpath('$workspaceFile', '$projectBaseDir')")
+                        fauxpas -o json check $projectName --workspace $workspaceRelativePath --scheme $scheme > sonar-reports/$(basename $projectName .xcodeproj)-$scheme-fauxpas.json
+                    fi
+                fi
+
+                done
+
+            done < tmpFileRunSonarSh
+            rm -rf tmpFileRunSonarSh
+
+	    fi
 
     else
         echo 'Skipping FauxPas (not installed)'
